Resolved Issue 178 (https://github.com/alimanfoo/petl/issues/178)
Prior method created a defaultdict, where key = the field name.  This posed problems when tables included duplicate field names in the header row.

The method now relies on the index of the field in the default dict, but formats and displays the field name when displayed.

For example:
from petl import see

table = [['foo', 'bar', 'foo'], ['a', '1', 'b'], ['c', 2, 'd']]

print(see(table))