avoid doctest fail when missing bcolz or pytables (hard setup)
This is just a quality of life improvement.
Changed docstring for skipping doctests for bcolz and pytables.
It's hard to setup bcolz and pytables because they need local packages installed.
Every time that included code in docstrings they failed leading and one need to erase the docstring for getting the test to pass.