Fix todataframe() to do not iterate the table multiple times
Converting the table data to a list simply with list() causes the
table to be materialized three times because list() calls __len__()
on the table twice to populate the list.

Anyhow, there is no need to have the table data in a list when an
iteratotor can be passed to pandas.DataFrame.from_records().

Fixes #578.